requirements.txt
langchain-google-genai
langchain
pypdf
chromadb
streamlit
python-dotenv
langchain-community